Reworked the local development environment chapter
A good chunk of the start of this chapter actually works better as the
introduction to the book. Given that, I created the introduction chapter
with it, and have reworked the start of the local dev environment
chapter to be more focused on just that, instead of a mixture of an
introduction and covering how to setup a local dev environment.