Update the local development environment chapter
This change updates the start of the local development environment
chapter to refine how the Docker environment is introduced and
explained. I wanted to better set out how it will work, so that each of
its parts makes sense, and that there is no dissonance generated. I also
added in some notes for what makes sense to also be covered, and added
links to relevant Docker documentation, so that the reader can quickly
find out more, as and when they want to.