Update the conclusion
This change changes the recap to make it clearer and more succinct, so
that it both holds me to account and the reader can know, whether or not
they've gotten what they were expecting from the introduction. Secondly,
it tightens the language and removes references to the original post,
replacing them with references to "short guide".